requirements.txt
termcolor>=2.4.0
pyyaml>=6.0.2
